refactor Game into CLI package
